+++ src/core/BUILD
@@ -1213,6 +1213,7 @@ grpc_cc_library(
         "//:gpr",
         "//:grpc_base",
         "//:handshaker",
+        "//:iomgr",
         "//:parse_address",
         "//:ref_counted_ptr",
         "//:uri_parser",
@@ -3286,6 +3287,7 @@ grpc_cc_library(
         "//:grpc_credentials_util",
         "//:grpc_security_base",
         "//:grpc_trace",
+        "//:iomgr",
         "//:parse_address",
         "//:promise",
         "//:ref_counted_ptr",
@@ -3360,8 +3362,10 @@ grpc_cc_library(
         "//:grpc_base",
         "//:grpc_security_base",
         "//:handshaker",
+        "//:iomgr",
         "//:promise",
         "//:ref_counted_ptr",
+        "//:resource_quota_api",
         "//:tsi_base",
         "//:tsi_fake_credentials",
     ],
@@ -3396,6 +3400,7 @@ grpc_cc_library(
         "//:grpc_base",
         "//:grpc_security_base",
         "//:handshaker",
+        "//:iomgr",
         "//:promise",
         "//:ref_counted_ptr",
         "//:tsi_base",
@@ -3454,6 +3459,7 @@ grpc_cc_library(
         "//:grpc_client_channel",
         "//:grpc_security_base",
         "//:handshaker",
+        "//:iomgr",
         "//:parse_address",
         "//:promise",
         "//:ref_counted_ptr",
@@ -3496,6 +3502,7 @@ grpc_cc_library(
         "//:grpc_security_base",
         "//:grpc_trace",
         "//:handshaker",
+        "//:iomgr",
         "//:promise",
         "//:ref_counted_ptr",
         "//:tsi_base",
@@ -3551,6 +3558,7 @@ grpc_cc_library(
         "//:grpc_security_base",
         "//:grpc_trace",
         "//:httpcli",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:uri_parser",
@@ -3622,6 +3630,7 @@ grpc_cc_library(
         "//:grpc_security_base",
         "//:grpc_trace",
         "//:handshaker",
+        "//:iomgr",
         "//:promise",
         "//:ref_counted_ptr",
         "//:tsi_base",
@@ -3703,6 +3712,7 @@ grpc_cc_library(
         "//:grpc_security_base",
         "//:grpc_trace",
         "//:httpcli",
+        "//:iomgr",
         "//:orphanable",
         "//:promise",
         "//:ref_counted_ptr",
@@ -3755,6 +3765,7 @@ grpc_cc_library(
         "//:grpc_credentials_util",
         "//:grpc_security_base",
         "//:httpcli",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:uri_parser",
@@ -3789,6 +3800,7 @@ grpc_cc_library(
         "//:grpc_base",
         "//:grpc_security_base",
         "//:handshaker",
+        "//:iomgr",
         "//:promise",
         "//:ref_counted_ptr",
         "//:tsi_base",
@@ -4057,6 +4069,7 @@ grpc_cc_library(
         "metadata_batch",
         "status_helper",
         "time",
+        "//:call_combiner",
         "//:channel_arg_names",
         "//:config",
         "//:debug_location",
@@ -4133,6 +4146,7 @@ grpc_cc_library(
         "slice",
         "slice_buffer",
         "validation_errors",
+        "//:call_trace",
         "//:channel_arg_names",
         "//:config",
         "//:gpr",
@@ -4661,6 +4675,7 @@ grpc_cc_library(
         "//:grpc_public_hdrs",
         "//:grpc_security_base",
         "//:grpc_trace",
+        "//:iomgr",
         "//:iomgr_timer",
         "//:orphanable",
         "//:parse_address",
@@ -4738,6 +4753,7 @@ grpc_cc_library(
         "//:grpc_security_base",
         "//:grpc_service_config_impl",
         "//:grpc_trace",
+        "//:iomgr",
         "//:parse_address",
         "//:ref_counted_ptr",
         "//:sockaddr_utils",
@@ -5713,6 +5729,7 @@ grpc_cc_library(
         "status_helper",
         "//:gpr_platform",
         "//:grpc_base",
+        "//:iomgr",
     ],
 )
 
@@ -5753,6 +5770,7 @@ grpc_cc_library(
         "//:grpc_resolver",
         "//:grpc_service_config_impl",
         "//:grpc_trace",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:uri_parser",
@@ -5811,6 +5829,7 @@ grpc_cc_library(
         "//:grpc_base",
         "//:grpc_resolver",
         "//:grpc_trace",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:uri_parser",
@@ -6008,6 +6027,7 @@ grpc_cc_library(
         "//:gpr",
         "//:grpc_base",
         "//:grpc_resolver",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:uri_parser",
@@ -6268,6 +6288,7 @@ grpc_cc_library(
         "//:grpc_trace",
         "//:grpc_transport_chttp2",
         "//:handshaker",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:sockaddr_utils",
@@ -6319,6 +6340,7 @@ grpc_cc_library(
         "//:grpc_trace",
         "//:grpc_transport_chttp2",
         "//:handshaker",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:sockaddr_utils",
@@ -6369,6 +6391,7 @@ grpc_cc_library(
         "//:grpc_base",
         "//:grpc_public_hdrs",
         "//:grpc_trace",
+        "//:iomgr",
         "//:promise",
         "//:ref_counted_ptr",
     ],
@@ -6472,6 +6495,7 @@ grpc_cc_library(
         "//:grpc_security_base",
         "//:grpc_trace",
         "//:httpcli",
+        "//:iomgr",
         "//:orphanable",
         "//:ref_counted_ptr",
         "//:uri_parser",
@@ -6841,6 +6865,7 @@ grpc_cc_library(
     ],
     deps = [
         "1999",
+        "call_final_info",
         "for_each",
         "if",
         "latch",
