+++ src/core/lib/channel/channel_stack.cc
@@ -347,3 +347,10 @@ void grpc_channel_stack::InitServerCallSpine(
     elem->filter->init_call(elem, call);
   }
 }
+
+void grpc_call_log_op(const char* file, int line, gpr_log_severity severity,
+                      grpc_call_element* elem,
+                      grpc_transport_stream_op_batch* op) {
+  gpr_log(file, line, severity, "OP[%s:%p]: %s", elem->filter->name, elem,
+          grpc_transport_stream_op_batch_string(op, false).c_str());
+}
