+++ src/core/lib/surface/call_trace.h
@@ -17,14 +17,8 @@
 
 #include <grpc/support/port_platform.h>
 
-#include "src/core/lib/channel/channel_fwd.h"
 #include "src/core/lib/debug/trace.h"
 
 extern grpc_core::TraceFlag grpc_call_trace;
-
-namespace grpc_core {
-const grpc_channel_filter* PromiseTracingFilterFor(
-    const grpc_channel_filter* filter);
-}
 
 #endif  // GRPC_SRC_CORE_LIB_SURFACE_CALL_TRACE_H
