+++ src/core/lib/transport/transport_op_string.cc
@@ -27,7 +27,6 @@
 #include <grpc/support/log.h>
 
 #include "src/core/lib/channel/channel_fwd.h"
-#include "src/core/lib/channel/channel_stack.h"
 #include "src/core/lib/gprpp/orphanable.h"
 #include "src/core/lib/gprpp/status_helper.h"
 #include "src/core/lib/slice/slice_buffer.h"
@@ -146,11 +145,4 @@ std::string grpc_transport_op_string(grpc_transport_op* op) {
   }
 
   return out;
-}
-
-void grpc_call_log_op(const char* file, int line, gpr_log_severity severity,
-                      grpc_call_element* elem,
-                      grpc_transport_stream_op_batch* op) {
-  gpr_log(file, line, severity, "OP[%s:%p]: %s", elem->filter->name, elem,
-          grpc_transport_stream_op_batch_string(op, false).c_str());
 }
